Handle aliases in chart dependencies
It's valid to have multiple dependencies with alternate aliases which
currently are removed on chart generation.

One way to resolve this is just to operate directly on the dependencies
list instead of generating a map keyed by name, which is less efficient
but avoids the problem of removing valid dependencies that use aliases.

Fixes: #18